Code/requirements.txt
paddleocr
Pillow
python-pptx
Flask
psutil
numpy
Werkzeug
psutil
configparser
openai
